Increased line counters to 64bits (input file sliced) & factorized

+++ src/io/input_buffer.c
@@ -48,7 +48,7 @@ input_buffer_t* input_buffer_new(const uint64_t buffer_size) {
   input_buffer->buffer_size = 0;
   input_buffer->buffer_allocated = buffer_size;
   // Line Index
-  input_buffer->line_lengths = vector_new(INPUT_BUFFER_LINE_OFFSETS_INIT,uint32_t);
+  input_buffer->line_lengths = vector_new(INPUT_BUFFER_LINE_OFFSETS_INIT,uint64_t);
   // Return
   return input_buffer;
 }
@@ -73,24 +73,11 @@ void input_buffer_annotate_lines(input_buffer_t* const input_buffer) {
   const char* const end = sentinel + buffer_size;
   while ((sentinel = memchr(sentinel,EOL,end-sentinel))) {
     ++sentinel;
-    vector_insert(line_lengths,sentinel-last_sentinel,uint32_t);
+    vector_insert(line_lengths,sentinel-last_sentinel,uint64_t);
     last_sentinel = sentinel;
   }
-  // Insert the length of the remaining chars
-  vector_insert(line_lengths,end-last_sentinel,uint32_t);
-  //  // Implementation based on pure counting (better for short lines)
-  //  uint32_t i, offset = 0;
-  //  for (i=0;i<buffer_size;++i) {
-  //    const char current_char = buffer[i];
-  //    if (gem_expect_false(current_char==EOL)) {
-  //      vector_insert(line_lengths,offset+1,uint32_t);
-  //      offset = 0;
-  //    } else {
-  //      ++offset;
-  //    }
-  //  }
-  //  // Insert the length of the remaining chars
-  //  vector_insert(line_lengths,offset,uint32_t);
+  // Insert the length of the remaining chars (Possible zero length line)
+  vector_insert(line_lengths,end-last_sentinel,uint64_t); // But not a line (not '\n' found)
 }
 uint64_t input_buffer_get_num_lines(input_buffer_t* const input_buffer) {
   return vector_get_used(input_buffer->line_lengths)-1;
